refactor(supervisor-operations-service): Split operations.controller.ts into 5 sub-domains
The 1,099-line, 33-route operations.controller.ts (spanning 5 unrelated
sub-domains: call logs, supervisor events, inventory, gatherings,
training topics) is replaced by 5 focused controller+routes pairs:
callLogs, events, inventory, gatherings, trainingTopics. Each controller
keeps only request-handling logic via a createXController(service)
factory; each routes file owns its route/middleware declarations and
response Zod schemas.

operations.module.ts now creates a single shared DocumentedRouter via
createDocumentedRouter() and calls all 5 registerXRoutes(doc, service)
functions against it — required because createDocumentedRouter()
produces exactly one OpenAPI registry per call; 5 independently-created
registries would have silently produced 5 incomplete Swagger docs
instead of 1 complete one. operations.service.ts/repository.ts are
unchanged (out of scope) — the service layer stays one class covering
all 5 sub-domains.

No behavior change — verified against a correct develop-based baseline
(33 routes, 23 paths): route-for-route identical OpenAPI doc
before/after, lint clean, all 186 existing tests pass, and a live
smoke test hitting a route in every one of the 5 sub-domains, all
correctly 401ing and confirming the shared-registry wiring is correct.

This is the final service in the routes/controller split refactor —
all 17 services, 49 controllers now follow the routes/controller
convention.
